+++ app/src/main/java/com/orange/ods/app/ui/components/ComponentDetailScreen.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.app.ui.components
 
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
@@ -27,7 +26,6 @@ import com.orange.ods.app.ui.utilities.composable.DetailScreenHeader
 import com.orange.ods.compose.component.list.OdsListItem
 import com.orange.ods.compose.component.list.OdsListItemIcon
 import com.orange.ods.compose.component.list.OdsListItemIconType
-import com.orange.ods.compose.component.list.iconType
 
 @Composable
 fun ComponentDetailScreen(
@@ -70,10 +68,9 @@ fun ComponentDetailScreen(
 @Composable
 private fun ComponentDetailLinkItem(label: String, composableName: String?, onClick: () -> Unit) {
     OdsListItem(
-        icon = { OdsListItemIcon(painter = painterResource(id = R.drawable.ic_play_outline), contentDescription = null) },
+        icon = OdsListItemIcon(OdsListItemIconType.Icon, painterResource(id = R.drawable.ic_play_outline), ""),
         text = label,
         secondaryText = composableName,
-        modifier = Modifier
-            .iconType(OdsListItemIconType.Icon)
-            .clickable { onClick() })
+        onClick = onClick
+    )
 }