+++ app/src/main/java/com/orange/ods/app/ui/components/appbars/top/ComponentTopAppBar.kt
@@ -46,7 +46,7 @@ import com.orange.ods.compose.component.appbar.top.OdsTopAppBarNavigationIcon
 import com.orange.ods.compose.component.chip.OdsChoiceChip
 import com.orange.ods.compose.component.chip.OdsChoiceChipsFlowRow
 import com.orange.ods.compose.component.list.OdsListItem
-import com.orange.ods.compose.component.list.OdsSwitchTrailing
+import com.orange.ods.compose.component.list.OdsListItemTrailingSwitch
 import com.orange.ods.compose.text.OdsTextBody2
 import com.orange.ods.compose.text.OdsTextCaption
 import com.orange.ods.compose.theme.OdsTheme
@@ -204,9 +204,7 @@ private fun CustomizationBottomSheetContent(customizationState: TopAppBarCustomi
         }
         OdsListItem(
             text = stringResource(id = R.string.component_app_bars_top_element_navigation_icon),
-            trailing = OdsSwitchTrailing(
-                checked = navigationIconEnabled
-            )
+            trailing = OdsListItemTrailingSwitch(navigationIconEnabled.value, { navigationIconEnabled.value = it })
         )
         ComponentCountRow(
             modifier = Modifier.padding(start = dimensionResource(id = com.orange.ods.R.dimen.screen_horizontal_margin)),
@@ -219,10 +217,7 @@ private fun CustomizationBottomSheetContent(customizationState: TopAppBarCustomi
         )
         OdsListItem(
             text = stringResource(id = R.string.component_app_bars_top_element_overflow_menu),
-            trailing = OdsSwitchTrailing(
-                checked = overflowMenuEnabled,
-                enabled = isOverflowMenuSwitchEnabled
-            )
+            trailing = OdsListItemTrailingSwitch(overflowMenuEnabled.value, { overflowMenuEnabled.value = it }, isOverflowMenuSwitchEnabled)
         )
         if (isLarge) {
             Subtitle(textRes = R.string.component_element_title, horizontalPadding = true)