+++ app/src/main/java/com/orange/ods/app/ui/components/cards/ComponentCard.kt
@@ -26,7 +26,7 @@ import com.orange.ods.compose.component.card.OdsHorizontalCardImagePosition
 import com.orange.ods.compose.component.chip.OdsChoiceChip
 import com.orange.ods.compose.component.chip.OdsChoiceChipsFlowRow
 import com.orange.ods.compose.component.list.OdsListItem
-import com.orange.ods.compose.component.list.OdsSwitchTrailing
+import com.orange.ods.compose.component.list.OdsListItemTrailingSwitch
 
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
@@ -39,12 +39,12 @@ fun ComponentCard(variant: Variant) {
             bottomSheetContent = {
                 OdsListItem(
                     text = stringResource(id = R.string.component_card_clickable),
-                    trailing = OdsSwitchTrailing(checked = clickable)
+                    trailing = OdsListItemTrailingSwitch(clickable.value, { clickable.value = it })
                 )
                 if (variant == Variant.CardVerticalHeaderFirst) {
                     OdsListItem(
                         text = stringResource(id = R.string.component_element_thumbnail),
-                        trailing = OdsSwitchTrailing(checked = thumbnailChecked)
+                        trailing = OdsListItemTrailingSwitch(thumbnailChecked.value, { thumbnailChecked.value = it })
                     )
                 } else if (variant == Variant.CardHorizontal) {
                     Subtitle(textRes = R.string.component_card_horizontal_image_position, horizontalPadding = true)
@@ -53,19 +53,25 @@ fun ComponentCard(variant: Variant) {
                         onValueChange = { value -> imagePosition.value = value },
                         modifier = Modifier.padding(horizontal = dimensionResource(id = com.orange.ods.R.dimen.spacing_m)),
                         chips = listOf(
-                            OdsChoiceChip(text = stringResource(id =  R.string.component_card_horizontal_image_position_start), value = OdsHorizontalCardImagePosition.Start),
-                            OdsChoiceChip(text = stringResource(id = R.string.component_card_horizontal_image_position_end), value = OdsHorizontalCardImagePosition.End)
+                            OdsChoiceChip(
+                                text = stringResource(id = R.string.component_card_horizontal_image_position_start),
+                                value = OdsHorizontalCardImagePosition.Start
+                            ),
+                            OdsChoiceChip(
+                                text = stringResource(id = R.string.component_card_horizontal_image_position_end),
+                                value = OdsHorizontalCardImagePosition.End
+                            )
                         )
                     )
                 }
                 OdsListItem(
                     text = stringResource(id = R.string.component_element_subtitle),
-                    trailing = OdsSwitchTrailing(checked = subtitleChecked)
+                    trailing = OdsListItemTrailingSwitch(subtitleChecked.value, { subtitleChecked.value = it })
                 )
                 if (variant in listOf(Variant.CardVerticalHeaderFirst, Variant.CardVerticalImageFirst, Variant.CardHorizontal)) {
                     OdsListItem(
                         text = stringResource(id = R.string.component_element_text),
-                        trailing = OdsSwitchTrailing(checked = textChecked)
+                        trailing = OdsListItemTrailingSwitch(textChecked.value, { textChecked.value = it })
                     )
                     ComponentCountRow(
                         title = stringResource(id = R.string.component_card_action_button_count),
@@ -81,7 +87,7 @@ fun ComponentCard(variant: Variant) {
                     if (!hasFirstButton) dividerChecked.value = false
                     OdsListItem(
                         text = stringResource(id = R.string.component_element_divider),
-                        trailing = OdsSwitchTrailing(checked = dividerChecked, enabled = hasFirstButton)
+                        trailing = OdsListItemTrailingSwitch(dividerChecked.value, { dividerChecked.value = it }, hasFirstButton)
                     )
                 }
             }) {