+++ app/src/main/java/com/orange/ods/app/ui/components/floatingactionbuttons/ComponentFloatingActionButton.kt
@@ -38,7 +38,7 @@ import com.orange.ods.compose.component.button.OdsFloatingActionButtonIcon
 import com.orange.ods.compose.component.chip.OdsChoiceChip
 import com.orange.ods.compose.component.chip.OdsChoiceChipsFlowRow
 import com.orange.ods.compose.component.list.OdsListItem
-import com.orange.ods.compose.component.list.OdsSwitchTrailing
+import com.orange.ods.compose.component.list.OdsListItemTrailingSwitch
 
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
@@ -106,11 +106,11 @@ fun ComponentFloatingActionButton() {
                 )
                 OdsListItem(
                     text = stringResource(id = R.string.component_element_text),
-                    trailing = OdsSwitchTrailing(checked = text, enabled = isTextEnabled)
+                    trailing = OdsListItemTrailingSwitch(text.value, { text.value = it }, isTextEnabled)
                 )
                 OdsListItem(
                     text = stringResource(id = R.string.component_floating_action_button_full_screen_width),
-                    trailing = OdsSwitchTrailing(checked = fullScreenWidth, enabled = isFullScreenWidthEnabled)
+                    trailing = OdsListItemTrailingSwitch(fullScreenWidth.value, { fullScreenWidth.value = it }, isFullScreenWidthEnabled)
                 )
             }) {
             Column(modifier = Modifier.verticalScroll(rememberScrollState())) {