+++ app/src/main/java/com/orange/ods/app/ui/components/utilities/ComponentCustomizationBottomSheet.kt
@@ -21,8 +21,10 @@ import androidx.compose.material.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.semantics
@@ -31,7 +33,7 @@ import androidx.compose.ui.unit.dp
 import com.orange.ods.app.R
 import com.orange.ods.app.ui.utilities.composable.OnResumeEffect
 import com.orange.ods.compose.component.bottomsheet.OdsBottomSheetScaffold
-import com.orange.ods.compose.component.list.OdsListItem
+import com.orange.ods.compose.text.OdsTextSubtitle1
 import com.orange.ods.compose.theme.OdsTheme
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
@@ -65,7 +67,7 @@ fun ComponentCustomizationBottomSheetScaffold(
         snackbarHost = snackbarHost,
         sheetPeekHeight = 56.dp,
         sheetContent = {
-            OdsListItem(
+            Row(
                 modifier = Modifier
                     .clickable {
                         coroutineScope.launch {
@@ -78,18 +80,25 @@ fun ComponentCustomizationBottomSheetScaffold(
                     }
                     .semantics {
                         stateDescription = bottomSheetHeaderStateDescription
-                    },
-                text = stringResource(id = titleResId),
-                icon = {
-                    val degrees = if (bottomSheetScaffoldState.bottomSheetState.isExpanded) 0f else -180f
-                    val angle by animateFloatAsState(targetValue = degrees)
-                    Icon(
-                        modifier = Modifier.rotate(angle),
-                        painter = painterResource(id = R.drawable.ic_chevron_down),
-                        contentDescription = null,
-                        tint = OdsTheme.colors.onSurface
-                    )
-                })
+                    }
+                    .fillMaxWidth()
+                    .height(56.dp)
+                    .padding(horizontal = dimensionResource(id = com.orange.ods.R.dimen.screen_horizontal_margin)),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                val degrees = if (bottomSheetScaffoldState.bottomSheetState.isExpanded) 0f else -180f
+                val angle by animateFloatAsState(targetValue = degrees, label = "ComponentCustomizationBottomSheetScaffoldIconRotation")
+                Icon(
+                    modifier = Modifier.rotate(angle),
+                    painter = painterResource(id = R.drawable.ic_chevron_down),
+                    contentDescription = null,
+                    tint = OdsTheme.colors.onSurface
+                )
+                OdsTextSubtitle1(
+                    modifier = Modifier.padding(start = dimensionResource(id = com.orange.ods.R.dimen.spacing_m)),
+                    text = stringResource(id = titleResId)
+                )
+            }
 
             Column(modifier = Modifier.verticalScroll(rememberScrollState())) {
                 bottomSheetContent()