+++ app/src/main/java/com/orange/ods/app/ui/guidelines/spacing/GuidelineSpacingScreen.kt
@@ -10,35 +10,37 @@
 
 package com.orange.ods.app.ui.guidelines.spacing
 
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.drawscope.DrawScope
+import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.orange.ods.compose.component.list.OdsCaptionTrailing
-import com.orange.ods.compose.component.list.OdsListItem
-import com.orange.ods.compose.component.list.divider
-import com.orange.ods.compose.text.OdsTextSubtitle1
 import com.orange.ods.app.R
-import com.orange.ods.app.ui.LocalThemeManager
 import com.orange.ods.app.ui.LocalMainTopAppBarManager
+import com.orange.ods.app.ui.LocalThemeManager
 import com.orange.ods.app.ui.guidelines.Guideline
 import com.orange.ods.app.ui.utilities.DrawableManager
 import com.orange.ods.app.ui.utilities.composable.DetailScreenHeader
 import com.orange.ods.app.ui.utilities.extension.isOrange
+import com.orange.ods.compose.component.list.OdsListItem
+import com.orange.ods.compose.component.list.OdsListItemIcon
+import com.orange.ods.compose.component.list.OdsListItemIconType
+import com.orange.ods.compose.component.list.OdsListItemTrailingCaption
+import com.orange.ods.compose.text.OdsTextSubtitle1
 import java.text.DecimalFormat
 import java.text.DecimalFormatSymbols
-import java.util.*
+import java.util.Locale
 
 private val ratioFormatter = DecimalFormat("0.#", DecimalFormatSymbols(Locale.ENGLISH))
 
@@ -63,55 +65,57 @@ fun GuidelineSpacingScreen() {
             )
         }
         items(Spacing.values()) { spacing ->
-            val dividerStartIndent =
-                dimensionResource(id = R.dimen.guideline_spacing_image_width) + dimensionResource(id = com.orange.ods.R.dimen.spacing_m).times(2)
             val dp = spacing.getDp()
             val ratio = spacing.getRatio()
             OdsListItem(
-                modifier = Modifier.divider(startIndent = dividerStartIndent),
                 text = spacing.tokenName,
                 secondaryText = stringResource(id = R.string.guideline_spacing_dp, dp.value.toInt()) + "\n",
                 singleLineSecondaryText = false,
-                icon = { GuidelineSpacingImage(spacing = spacing) },
-                trailing = OdsCaptionTrailing(
-                    text = stringResource(
+                icon = OdsListItemIcon(OdsListItemIconType.SquareImage, rememberGuidelineSpacingPainter(spacing = spacing), ""),
+                trailing = OdsListItemTrailingCaption(
+                    stringResource(
                         id = R.string.guideline_spacing_ratio,
                         if (ratio == 0.0f) "-" else ratioFormatter.format(ratio)
                     )
-                )
+                ),
+                divider = true
             )
         }
     }
 }
 
 @Composable
-fun GuidelineSpacingImage(spacing: Spacing) {
-    // Spacing width is at least 1 dp to make spacing-none visible
-    val spacingWidth = dimensionResource(id = spacing.dimenRes).coerceAtLeast(1.dp)
-    val imageWidth = dimensionResource(id = R.dimen.guideline_spacing_image_width)
-    val imageHeight = dimensionResource(id = R.dimen.guideline_spacing_image_height)
-    val isOrangeTheme = LocalThemeManager.current.currentThemeConfiguration.isOrange
+private fun rememberGuidelineSpacingPainter(spacing: Spacing): GuidelineSpacingPainter {
+    with(LocalDensity.current) {
+        val width = dimensionResource(id = com.orange.ods.R.dimen.list_wide_image_width).toPx()
+        val height = dimensionResource(id = com.orange.ods.R.dimen.list_wide_image_height).toPx()
+        val spacingWidth = dimensionResource(id = spacing.dimenRes).coerceAtLeast(1.dp).toPx() // Spacing width is at least 1 dp to make spacing-none visible
+        val isOrangeTheme = LocalThemeManager.current.currentThemeConfiguration.isOrange
+        val spacingColor = if (isOrangeTheme) 0xff4bb4e6 else 0xff949494
 
-    Canvas(
-        modifier = Modifier
-            .width(imageWidth)
-            .height(imageHeight),
-    ) {
+        return remember(spacing) {
+            GuidelineSpacingPainter(Size(width, height), spacingWidth, spacingColor)
+        }
+    }
+}
+
+private class GuidelineSpacingPainter(override val intrinsicSize: Size, val spacingWidth: Float, val spacingColor: Long) : Painter() {
+
+    override fun DrawScope.onDraw() {
         // Background
         drawRect(Color(0xfff2f2f7))
         // Banner
         val bannerHeight = 16.dp
         drawRect(
             Color(0xff595959),
-            Offset(0.0f, ((imageHeight - bannerHeight) / 2.0f).toPx()),
-            Size(imageWidth.toPx(), bannerHeight.toPx())
+            Offset(0.0f, ((intrinsicSize.height - bannerHeight.toPx()) / 2.0f)),
+            Size(intrinsicSize.width, bannerHeight.toPx())
         )
         // Spacing
-        val spacingColor = if (isOrangeTheme) 0xff4bb4e6 else 0xff949494
         drawRect(
             Color(spacingColor),
-            Offset(((imageWidth - spacingWidth) / 2.0f).toPx(), 0.0f),
-            Size(spacingWidth.toPx(), imageHeight.toPx())
+            Offset(((intrinsicSize.width - spacingWidth) / 2.0f), 0.0f),
+            Size(spacingWidth, intrinsicSize.height)
         )
     }
 }