+++ lib/src/main/java/com/orange/ods/compose/component/chip/OdsChoiceChipsFlowRow.kt
@@ -17,10 +17,8 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.selection.selectableGroup
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.semantics.SemanticsPropertyReceiver
@@ -53,8 +51,6 @@ fun <T> OdsChoiceChipsFlowRow(
     onValueChange: (value: T) -> Unit,
     modifier: Modifier = Modifier
 ) {
-    var selectedChipValue by remember { mutableStateOf(value) }
-
     FlowRow(
         modifier = modifier
             .fillMaxWidth()
@@ -62,9 +58,8 @@ fun <T> OdsChoiceChipsFlowRow(
         horizontalArrangement = Arrangement.spacedBy(dimensionResource(id = R.dimen.spacing_s)),
         content = {
             chips.forEachIndexed { index, odsChoiceChip ->
-                odsChoiceChip.Content(selected = selectedChipValue == odsChoiceChip.value) { selected ->
+                odsChoiceChip.Content(selected = value == odsChoiceChip.value) { selected ->
                     if (selected) {
-                        selectedChipValue = chips[index].value
                         onValueChange(chips[index].value)
                     }
                 }